Add valid start and end date check
This change adds a simple check to make sure that the isCurrent
method of DateUtils will successfully return true when the
startDate is before the current time and the current time is
before the endDate. As the code was already completed as part
of other tests, this code is only a test change.