Feature/current time (#1)
* Add valid start date check

This change adds validation code to DateUtils so that if the start date
is after the current date, then it returns false signifying that the
start date hasn't happened yet.

* Add valid end date check

This change adds an end date parameter to the isCurrent method in the
DateUtils class and with this parameter we can test that a start and
end date envelope the current date, if they do not then the test fails
and false is returned.

* Refactor DateUtilsTest to use data tables

This is required to de-dupe the test cases and to make it easier
to read for future developers.

* Add valid start and end date check

This change adds a simple check to make sure that the isCurrent
method of DateUtils will successfully return true when the
startDate is before the current time and the current time is
before the endDate. As the code was already completed as part
of other tests, this code is only a test change.